Create User Account.php
I can insert into both user and account tables, but the location for user is only the city. I still need to figure out how to concatenate the State with the city. I can also change the input method for state to take a text instead of a scroll bar